requirements.txt
fastapi
uvicorn
pandas
scikit-surprise
joblib